cgo option build
With this patch, we've made the external C parser optional for generateOutput. As cgo can be problematic for some users, we've added an option to bypass it and test generateOutput against a sample dot file instead. This change should make the tool more accessible and user-friendly.

Suggested-by: Maurizio Papini <mpapini@redhat.com>
