--- libdotparser.so
@@ -1,1 +1,0 @@
-dot_parser/libdotparser.so
